Refactor import statements and remove unused trade logic code
